server/requirements.txt
python-dotenv
openai
together
fastapi
transformers[torch]
hydra-core
omegaconf
bitarray
scikit-learn
sacrebleu